-Implemented new function to create Matrix, matrix() -Implemented new function to create Tuple, tuple() -Implemented matrix related operations, multiplication and transpose -Implemented matrix helper functions, compare function | free function and print function

+++ notes/complex_opts.c
@@ -49,9 +49,9 @@ double	magnitude(double *vector, int size)
 	return (sqrt(mag));
 }
 
-int	dot(double *vector1, double *vector2, int size)
+double	dot(double *vector1, double *vector2, int size)
 {
-	int	product;
+	double	product;
 	int	i;
 
 	product = 0;
